Move checkCountry to GoldleRunCrew
